add feature flag and restore original work mode
