+++ scripts/babel/transform-prevent-infinite-loops.js
@@ -13,10 +13,10 @@
 // This should be reasonable for all loops in the source.
 // Note that if the numbers are too large, the tests will take too long to fail
 // for this to be useful (each individual test case might hit an infinite loop).
-const MAX_SOURCE_ITERATIONS = 10;
+const MAX_SOURCE_ITERATIONS = 15000;
 // Code in tests themselves is permitted to run longer.
 // For example, in the fuzz tester.
-const MAX_TEST_ITERATIONS = 10;
+const MAX_TEST_ITERATIONS = 50000;
 
 module.exports = ({types: t, template}) => {
   // We set a global so that we can later fail the test
